feat(auth): move OpenIdProvider and OpenIdCertificate types and impl
