+++ src/observatory/src/impls.rs
@@ -4,24 +4,16 @@ use crate::templates::{
     FAILED_DEPOSIT_CYCLES_TXT,
 };
 use crate::types::interface::NotifyStatus;
-use crate::types::state::{
-    HeapState, Notification, NotificationKey, NotificationStatus, OpenIdCertificate,
-    OpenIdProvider, State,
-};
+use crate::types::state::{HeapState, Notification, NotificationKey, NotificationStatus, State};
 use ic_cdk::api::time;
 use ic_stable_structures::storable::Bound;
 use ic_stable_structures::Storable;
-use junobuild_auth::openid::jwt::types::cert::Jwks;
 use junobuild_shared::serializers::{
     deserialize_from_bytes, serialize_into_bytes, serialize_to_bytes,
 };
 use junobuild_shared::types::interface::NotifyArgs;
-use junobuild_shared::types::state::{
-    NotificationKind, SegmentKind, Timestamp, Version, Versioned,
-};
-use junobuild_shared::version::next_version;
+use junobuild_shared::types::state::{NotificationKind, SegmentKind};
 use std::borrow::Cow;
-use std::fmt::{Display, Formatter, Result as FmtResult};
 use time::OffsetDateTime;
 
 impl Default for State {
@@ -243,66 +235,6 @@ impl NotifyStatus {
             pending,
             sent,
             failed,
-        }
-    }
-}
-
-impl OpenIdProvider {
-    pub fn jwks_url(&self) -> &'static str {
-        match self {
-            Self::Google => "https://www.googleapis.com/oauth2/v3/certs",
-        }
-    }
-}
-
-impl Versioned for OpenIdCertificate {
-    fn version(&self) -> Option<Version> {
-        self.version
-    }
-}
-
-impl OpenIdCertificate {
-    fn get_next_version(current_certificate: &Option<OpenIdCertificate>) -> Version {
-        next_version(current_certificate)
-    }
-
-    pub fn init(jwks: &Jwks, expires_at: &Option<Timestamp>) -> Self {
-        let now = time();
-
-        let version = Self::get_next_version(&None);
-
-        Self {
-            jwks: jwks.clone(),
-            expires_at: *expires_at,
-            created_at: now,
-            updated_at: now,
-            version: Some(version),
-        }
-    }
-
-    pub fn update(
-        current_certificate: &OpenIdCertificate,
-        jwks: &Jwks,
-        expires_at: &Option<Timestamp>,
-    ) -> Self {
-        let now = time();
-
-        let version = Self::get_next_version(&Some(current_certificate.clone()));
-
-        Self {
-            jwks: jwks.clone(),
-            expires_at: *expires_at,
-            updated_at: now,
-            version: Some(version),
-            ..current_certificate.clone()
-        }
-    }
-}
-
-impl Display for OpenIdProvider {
-    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-        match self {
-            OpenIdProvider::Google => write!(f, "Google"),
         }
     }
 }
