+++ src/observatory/src/lib.rs
@@ -16,8 +16,8 @@ use crate::types::interface::GetNotifications;
 use crate::types::interface::GetOpenIdCertificate;
 use crate::types::interface::NotifyStatus;
 use crate::types::state::Env;
-use crate::types::state::OpenIdCertificate;
 use ic_cdk_macros::export_candid;
+use junobuild_auth::openid::types::provider::OpenIdCertificate;
 use junobuild_shared::types::interface::NotifyArgs;
 use junobuild_shared::types::interface::{DeleteControllersArgs, SetControllersArgs};
 use junobuild_shared::types::state::Controllers;
